+++ CHANGES.md
@@ -17,6 +17,7 @@
 * Refactored code dealing with requests and authentication to the `STACpopulator/requests.py` file
 * Add `--log-file` command line option to specify a non-default location to write log files to
 * Add `cordex6` extension and `CORDEX-CMIP6_Ouranos` implementation. This includes a refactoring of base extension classes.
+* Add an `xscen` extension demonstrating how to add properties to a STAC Item.
 
 ## [0.6.0](https://github.com/crim-ca/stac-populator/tree/0.6.0) (2024-02-22)